fixup! Cover ai.koog.agents.core.agent.context with tests
